+++ src/features/Conversation/components/ChatItem/utils.ts
@@ -1,26 +1,14 @@
-import {
-  ARTIFACT_TAG_REGEX,
-  ARTIFACT_THINKING_TAG_REGEX,
-  THINKING_TAG_REGEX,
-} from '@/const/plugin';
+import { ARTIFACT_TAG_REGEX, ARTIFACT_THINKING_TAG_REGEX } from '@/const/plugin';
 
 /**
  * Replace all line breaks in the matched `lobeArtifact` tag with an empty string
  */
 export const processWithArtifact = (input: string = '') => {
   let output = input;
-
-  const thinkMatch = THINKING_TAG_REGEX.exec(input);
-
-  // If the input contains the `think` tag, replace all line breaks with an empty string
-  if (thinkMatch)
-    output = input.replace(ARTIFACT_THINKING_TAG_REGEX, (match) =>
-      match.replaceAll(/\r?\n|\r/g, ''),
-    );
-  const lobeThinkMatch = ARTIFACT_THINKING_TAG_REGEX.exec(input);
+  const thinkMatch = ARTIFACT_THINKING_TAG_REGEX.exec(input);
 
   // If the input contains the `lobeThinking` tag, replace all line breaks with an empty string
-  if (lobeThinkMatch)
+  if (thinkMatch)
     output = input.replace(ARTIFACT_THINKING_TAG_REGEX, (match) =>
       match.replaceAll(/\r?\n|\r/g, ''),
     );
