--- src/features/Conversation/components/MarkdownElements/Thinking/index.ts
@@ -1,10 +1,0 @@
-import createOneLineRehypePlugin from '../rehypePlugin/createOneLineRehypePlugin';
-import Component from './Render';
-
-const ThinkingElement = {
-  Component,
-  rehypePlugin: createOneLineRehypePlugin('think'),
-  tag: 'think',
-};
-
-export default ThinkingElement;
