+++ src/features/Conversation/components/MarkdownElements/index.ts
@@ -1,5 +1,4 @@
 import LobeArtifact from './LobeArtifact';
 import LobeThinking from './LobeThinking';
-import Thinking from './Thinking';
 
-export const markdownElements = [Thinking, LobeArtifact, LobeThinking];
+export const markdownElements = [LobeArtifact, LobeThinking];
